+++ kolena/workflow/_datatypes.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 import dataclasses
 import json
-import mimetypes
 from abc import ABCMeta
 from abc import abstractmethod
 from collections import OrderedDict
@@ -239,24 +238,6 @@ class DataObject(metaclass=ABCMeta):
         return self._to_dict()
 
 
-TEST_SAMPLE_TYPE = "TEST_SAMPLE"
-
-_DATAPOINT_TYPE_MAP = {
-    "image": f"{TEST_SAMPLE_TYPE}/IMAGE",
-    "application/pdf": f"{TEST_SAMPLE_TYPE}/DOCUMENT",
-    "text": f"{TEST_SAMPLE_TYPE}/DOCUMENT",
-    "video": f"{TEST_SAMPLE_TYPE}/VIDEO",
-}
-
-_TYPED_DATAOBJECT_MAP = {}
-
-
-def _set_map() -> None:
-    from kolena.workflow.annotation import BoundingBox
-
-    _TYPED_DATAOBJECT_MAP["ANNOTATION/BOUNDING_BOX"] = BoundingBox
-
-
 def _serialize_dataobject(x: Any) -> Any:
     if isinstance(x, list):
         return [item._to_dict() if isinstance(item, DataObject) else item for item in x]
@@ -271,7 +252,7 @@ def _deserialize_dataobject(x: Any) -> Any:
     if isinstance(x, dict) and DATA_TYPE_FIELD in x:
         data = {**x}
         data_type = data.pop(DATA_TYPE_FIELD)
-        typed_dataobject = _TYPED_DATAOBJECT_MAP.get(data_type, None)
+        typed_dataobject = _DATA_TYPE_MAP.get(data_type, None)
         if typed_dataobject:
             return typed_dataobject._from_dict(data)
 
@@ -280,24 +261,6 @@ def _deserialize_dataobject(x: Any) -> Any:
 
 _serialize_series = np.vectorize(_serialize_dataobject)
 _deserialize_series = np.vectorize(_deserialize_dataobject)
-
-
-def _dataobject_type(obj: "TypedDataObject") -> str:
-    obj_type = obj._data_type()
-    return f"{obj_type._data_category()}/{obj_type.value}"
-
-
-def _infer_datatype(x: str) -> str:
-    mtype, _ = mimetypes.guess_type(x)
-    if mtype:
-        main_type, sub_type = mtype.split("/")
-        datatype = _DATAPOINT_TYPE_MAP.get(mtype, None) or _DATAPOINT_TYPE_MAP.get(main_type, None)
-        if datatype is not None:
-            return datatype
-    elif x.endswith(".pcd"):
-        return f"{TEST_SAMPLE_TYPE}/POINT_CLOUD"
-
-    return f"{TEST_SAMPLE_TYPE}/CUSTOM"
 
 
 def _serialize_json(x: Any) -> Any:
@@ -326,7 +289,6 @@ def dataframe_to_csv(df: pd.DataFrame, *args, **kwargs) -> Union[str, None]:
 
 
 def dataframe_from_csv(*args, **kwargs) -> pd.DataFrame:
-    _set_map()
     df = pd.read_csv(*args, **kwargs)
     columns = list(df.select_dtypes(include="object").columns)
     df_post = df.select_dtypes(exclude="object")
@@ -337,7 +299,6 @@ def dataframe_from_csv(*args, **kwargs) -> pd.DataFrame:
 
 
 def dataframe_from_json(*args, **kwargs) -> pd.DataFrame:
-    _set_map()
     df = pd.read_json(*args, **kwargs)
     columns = list(df.select_dtypes(include="object").columns)
     df_post = df.select_dtypes(exclude="object")
